Add files via upload
1. Added decompose_unit function to handle units that sympy doesn't accept. This helps kamodo accept functions with arg_units or return units, but breaks later in to_latex representation of the function.
2. Added tests in test_kamodo.py related to this (bottom of file). Not sure how to test for proper latex representation, though.
3. Replaced getfullargspec(func).args line in get_function_args b/c was returning an empty argument list for wrapped functions with dimensionless input and output units. Now works with plasmapy wrapper.
4. Removed nT and nPa definitions in util.py because it is included in the prefix code directly below the definitions. The same units still work and are represented correctly.
5. Added a note about the N and S puzzle in util.py

Note: sympy_units has newtons and siemens (sympy_units.N and sympy_units.S), but won't recognize the units through kamodo. Asher mentioned something about putting them through a symbol call later that may be messing this up? The new decompose_unit feature gets around this for now, but the units in the kamodofy call have to be corrected. A similar approach works in the kamodofyplasmapy wrapper without any error in the latex representation of the corresponding kamodo object, but the correction occurs in the units and arg_units definitions in the kamodofy call itself, not after the units and arg_units are defined.