New logic to conserve memory
Creates a list of interpolators, one for every time step, for any dataset with at least 3 dimensions. Does not store the converted array in memory. Currently tested to work with netcdf4 files. Need to test with h5 files. The new logic is not yet implemented in any other readers except for ctipe. The ctipe file converter now calculates the median km grid for each ilev grid and stores a time step from the earlier file (if available) for later interpolation between files.