HAPI-Kamodo scripts
HAPI_KamodoRealFlightJSONs: creates dataset JSON for a given dataset (the sat_name is ignored), and creates the parameter(s) JSON for a given set of parameters. 
HAPI_KamodoRealFlight: interpolates the chosen satellite trajectory through the chosen dataset and outputs the data into HAPI csv and binary formats (not tested with HAPI yet)
HAPI_KamodoJSONs: creates dataset JSON for a given dataset (the sat_name is ignored), and creates the parameter(s) JSON for a given set of parameters, INCLUDING the coordinate grids
CTIPe_*.json: examples of the JSONs created. (RR = RealFlight version for parameter output)